Se actualizan las entidades de usuarios(variable Long) y Proveedores(creado completo) DB

+++ FivewareBack/src/main/java/co/edu/unbosque/api/UsuariosApi.java
@@ -34,7 +34,7 @@ public class UsuariosApi {
 	}
 
 	@DeleteMapping("/eliminar/{id}")
-	public void eliminar(@PathVariable("id") Integer id) {
+	public void eliminar(@PathVariable("id") Long id) {
 		usuarioDao.deleteById(id);
 	}
