+++ FivewareBack/src/main/java/co/edu/unbosque/dao/UsuariosDAO.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.unbosque.model.Usuarios;
 
-public interface UsuariosDAO extends JpaRepository<Usuarios, Integer> {
+public interface UsuariosDAO extends JpaRepository<Usuarios, Long> {
 
 }
